refactor crud in Todo

+++ DataAccess/TodoRepository.cs
@@ -5,30 +5,30 @@ namespace DataAccess
     internal class TodoRepository(TodoDb context) : ITodoRepository
     {
 
-        public async Task<List<Todo>> GetAllAsync(CancellationToken cancellationToken = default)
+        public async Task<List<Todo>> GetAllAsync(Guid? accountId, CancellationToken cancellationToken = default)
         {
-            return await context.Todos.ToListAsync(cancellationToken);
+            return await context.Todos.Where(x => x.AccountId == accountId).ToListAsync();
         }
         
-        public async Task<List<Todo>> GetCompletedAsync(CancellationToken cancellationToken = default)
+        public async Task<List<Todo>> GetCompletedAsync(Guid? accountId, CancellationToken cancellationToken = default)
         {
-            return await context.Todos.Where(x => x.IsComplete).ToListAsync();
+            return await context.Todos.Where(x => x.AccountId == accountId && x.IsComplete).ToListAsync();
         }
 
-        public async Task<List<Todo>> GetUncompletedAsync(CancellationToken cancellationToken = default)
+        public async Task<List<Todo>> GetUncompletedAsync(Guid? accountId, CancellationToken cancellationToken = default)
         {
-            return await context.Todos.Where(x => !x.IsComplete).ToListAsync();
+            return await context.Todos.Where(x => x.AccountId == accountId && !x.IsComplete).ToListAsync();
         }
 
-        public async Task<Todo?> GetByIdAsync(Guid id, CancellationToken cancellationToken = default)
+        public async Task<Todo?> GetByIdAsync(Guid? accountId, Guid id, CancellationToken cancellationToken = default)
         {
-            return await context.Todos.FirstOrDefaultAsync(x => x.Id == id);
+            return await context.Todos.FirstOrDefaultAsync(x => x.AccountId == accountId && x.Id == id);
         }
 
-        public async Task<List<Todo>?> GetByNameAsync(string name, CancellationToken cancellationToken = default)
+        public async Task<List<Todo>?> GetByNameAsync(Guid? accountId, string name, CancellationToken cancellationToken = default)
         {
             return await context.Todos
-            .Where(t => t.Name != null && t.Name.ToLower().Contains(name.ToLower()))
+            .Where(x => x.AccountId == accountId && x.Name != null && x.Name.ToLower().Contains(name.ToLower()))
             .ToListAsync(cancellationToken);
         }
 
@@ -44,21 +44,31 @@ namespace DataAccess
             await context.SaveChangesAsync();
         }
 
-        public async Task UpdateByIdAsync(Todo todo, CancellationToken cancellationToken = default)
+        public async Task UpdateByIdAsync(Guid? accountId, Todo todo, CancellationToken cancellationToken = default)
         {
-            context.Todos.Update(todo);
-            await context.SaveChangesAsync(cancellationToken);
+            if (todo.AccountId == accountId)
+            {
+                context.Todos.Update(todo);
+                await context.SaveChangesAsync(cancellationToken);
+            }
 
         }
 
-        public async Task DeleteAsync(Todo todo, CancellationToken cancellationToken = default)
+        public async Task DeleteAsync(Guid? accountId, Todo todo, CancellationToken cancellationToken = default)
         {
-            context.Todos.Remove(todo);
-            await context.SaveChangesAsync(cancellationToken);
+            if (todo.AccountId == accountId)
+            {
+                context.Todos.Remove(todo);
+                await context.SaveChangesAsync(cancellationToken);
+            }
         }
 
-        public async Task DeleteBulkAsync(List<Todo> todos, CancellationToken cancellationToken = default)
+        public async Task DeleteBulkAsync(Guid? accountId, List<Todo> todos, CancellationToken cancellationToken = default)
         {
+            if (todos.Any(todo => todo.AccountId != accountId))
+            {
+                throw new UnauthorizedAccessException("Some Todos do not belong to the specified account.");
+            }
             context.Todos.RemoveRange(todos);
             await context.SaveChangesAsync(cancellationToken);
         }
